backtest/requirements.txt
backtrader
matplotlib
numpy
pandas
pymongo
python-binance